feat: migrate admin module
