+++ libs/frontend/modules/admin/src/index.ts
@@ -1,2 +1,3 @@
+export * from './use-cases/admin.endpoints'
 export * from './use-cases/reset-data/ResetDataButton'
 export * from './use-cases/seed-base-types/SeedBaseTypesButton'
