+++ libs/frontend/modules/admin/src/use-cases/admin.endpoints.graphql
@@ -1,3 +1,7 @@
+mutation ResetData {
+  resetData
+}
+
 mutation SeedBaseTypes {
   seedBaseTypes
 }
