+++ libs/frontend/modules/admin/src/use-cases/admin.endpoints.ts
@@ -1,0 +1,4 @@
+import { api as generatedApi } from './admin.endpoints.graphql.gen'
+
+export { generatedApi as adminEndpoints }
+export const { useResetDataMutation, useSeedBaseTypesMutation } = generatedApi
