--- libs/frontend/modules/admin/src/use-cases/reset-data/ResetData.web.graphql
@@ -1,3 +1,0 @@
-mutation ResetData {
-  resetData
-}
