+++ libs/frontend/modules/admin/src/use-cases/seed-base-types/SeedBaseTypesButton.tsx
@@ -1,5 +1,5 @@
 import { Button } from 'antd'
-import { useSeedBaseTypesMutation } from './SeedBaseTypes.web.graphql.gen'
+import { useSeedBaseTypesMutation } from '../admin.endpoints'
 
 export const SeedBaseTypesButton = () => {
   const [seedBaseTypes] = useSeedBaseTypesMutation()
